refactor: move wallet specific utils to wallet.ts file
