+++ src/shared/utils/index.ts
@@ -14,14 +14,7 @@ import {
 } from 'react-native-screenshot-prevent'
 import { useTranslation } from 'react-i18next'
 import { useIsFocused } from '@react-navigation/native'
-import Config from 'react-native-config'
-import { providers } from 'ethers'
-import { RifRelayConfig } from '@rsksmart/rif-relay-light-sdk'
 
-import { ChainID, EOAWallet, OnRequest, WalletState } from 'lib/eoaWallet'
-import { RelayWallet } from 'lib/relayWallet'
-
-import { Wallet } from '../wallet'
 import { ErrorWithMessage } from '../types'
 
 export const delay = (delayMs: number) => {
@@ -203,70 +196,4 @@ export const getFormattedTokenValue = (tokenValue: string) => {
     tokenValueArr[1].slice(0, 7).concat(...lastAfterZero) +
     ending
   )
-}
-
-export const createAppWallet = async (
-  mnemonic: string,
-  chainId: ChainID,
-  jsonRpcProvider: providers.StaticJsonRpcProvider,
-  onRequest: OnRequest,
-  config: RifRelayConfig,
-  cache?: (privateKey: string, mnemonic?: string) => void,
-) => {
-  const useRelay = Config.USE_RELAY === 'true'
-  console.log('USE RELAY createAppWallet', useRelay)
-  let wallet: Wallet
-
-  if (useRelay) {
-    wallet = await RelayWallet.create(
-      mnemonic,
-      chainId,
-      jsonRpcProvider,
-      onRequest,
-      config,
-      cache,
-    )
-  } else {
-    wallet = EOAWallet.create(
-      mnemonic,
-      chainId,
-      jsonRpcProvider,
-      onRequest,
-      cache,
-    )
-  }
-
-  return wallet
-}
-
-export const loadAppWallet = async (
-  keys: WalletState,
-  chainId: ChainID,
-  jsonRpcProvider: providers.StaticJsonRpcProvider,
-  onRequest: OnRequest,
-  config: RifRelayConfig,
-) => {
-  const useRelay = Config.USE_RELAY === 'true'
-  console.log('USE RELAY loadAppWallet', useRelay)
-
-  let wallet: Wallet
-
-  if (useRelay) {
-    wallet = await RelayWallet.fromWalletState(
-      keys,
-      chainId,
-      jsonRpcProvider,
-      onRequest,
-      config,
-    )
-  } else {
-    wallet = EOAWallet.fromWalletState(
-      keys,
-      chainId,
-      jsonRpcProvider,
-      onRequest,
-    )
-  }
-
-  return wallet
 }
